Some fixes
- Change getClusterIDMap function.
- Add the return after deletion logic.
- some minor fixes.
